adding cmake and python requirements file

+++ src/GraphStrategy.cpp
@@ -6,6 +6,7 @@
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
 #endif
+
 const float DEGREE = 360.0;
 
 #pragma region Path Calculation
@@ -29,7 +30,7 @@ std::vector<Point> GraphStrategy::calculate(const IFlight* context, const Point&
 	std::unordered_map<GraphStrategy::Node, bool> visitedNodes;
 	//track the nodes that have been popped
 	std::vector<Node*> popsNode;
-
+	
 	//define the start node
 	Node* cur = new Node(position, curAngel, 0, position.distance(destination));
 	queue.push(cur);
@@ -41,10 +42,10 @@ std::vector<Point> GraphStrategy::calculate(const IFlight* context, const Point&
 		queue.pop();
 		popsNode.push_back(cur);
 		visitedNodes[*cur] = false;
-
+		
 		// Goal check
 		if (cur->position.distance(destination) < stepSize) {
-			float targetAngle = angleToTarget(cur->position, destination);
+			double targetAngle = angleToTarget(cur->position, destination);
 			normalizeAngle(targetAngle);
 			//check its in the same direction
 			if (angleDifference(cur->angle, targetAngle) < maxAngle) {
